# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Calculadora de Débitos Judiciais",
    "version": "2.3.2"
  },
  "paths": {
...
    "/api/v2/calculos/padroes": {
      "get": {
        "tags": [
          "Cálculos"
        ],
        "summary": "Defaults",
        "description": "Recomenda competência sem ultrapassar a cobertura do índice selecionado.",
        "operationId": "defaults_api_v2_calculos_padroes_get",
        "parameters": [
          {
            "name": "indice",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Indice"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationDefaults"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/calculos/politica/{origem_calculo}": {
      "get": {
        "tags": [
          "Cálculos"
        ],
        "summary": "Calculation Policy",
        "description": "Expõe padrões e metadados oficiais sem duplicá-los no frontend.",
        "operationId": "calculation_policy_api_v2_calculos_politica__origem_calculo__get",
        "parameters": [
          {
            "name": "origem_calculo",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "manual",
                "processo"
              ],
              "type": "string",
              "title": "Origem Calculo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationPolicyView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/calculos/honorarios/preparar": {
      "post": {
        "tags": [
          "Cálculos"
        ],
        "summary": "Prepare Fees",
        "description": "Mostra os valores nominais para revisão antes da chamada ao motor.",
        "operationId": "prepare_fees_api_v2_calculos_honorarios_preparar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeePreparation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Installment-Output"
                  },
                  "type": "array",
                  "title": "Response Prepare Fees Api V2 Calculos Honorarios Preparar Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/calculos/historico": {
      "get": {
        "tags": [
          "Cálculos"
        ],
        "summary": "History",
        "description": "Lista cálculos paginados; as versões são carregadas somente após expansão.",
        "operationId": "history_api_v2_calculos_historico_get",
        "parameters": [
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Pagina"
            }
          },
          {
            "name": "tamanho_pagina",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Tamanho Pagina"
            }
          },
          {
            "name": "busca",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Busca"
            }
          },
          {
            "name": "origem",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "manual",
                    "processo"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origem"
            }
          },
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ativo",
                    "arquivado",
                    "cancelado"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estado"
            }
          },
          {
            "name": "indice",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Indice"
            }
          },
          {
            "name": "criado_por",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Criado Por"
            }
          },
          {
            "name": "atualizado_de",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Atualizado De"
            }
          },
          {
            "name": "atualizado_ate",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Atualizado Ate"
            }
          },
          {
            "name": "ordenacao",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "processo",
                "atualizado_desc",
                "atualizado_asc",
                "criado_desc",
                "criado_asc"
              ],
              "type": "string",
              "default": "processo",
              "title": "Ordenacao"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationHistoryPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/calculos/{calculo_id}/versoes": {
      "get": {
        "tags": [
          "Cálculos"
        ],
        "summary": "Versions",
        "description": "Lazy loading das versões de um cálculo já localizado no histórico.",
        "operationId": "versions_api_v2_calculos__calculo_id__versoes_get",
        "parameters": [
          {
            "name": "calculo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Calculo Id"
            }
          },
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Pagina"
            }
          },
          {
            "name": "tamanho_pagina",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Tamanho Pagina"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationVersionsPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/calculos/{calculo_id}/comparar": {
      "get": {
        "tags": [
          "Cálculos"
        ],
        "summary": "Compare Versions",
        "description": "Compara parâmetros, parcelas e impacto no total entre duas versões.",
        "operationId": "compare_versions_api_v2_calculos__calculo_id__comparar_get",
        "parameters": [
          {
            "name": "calculo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Calculo Id"
            }
          },
          {
            "name": "versao_origem",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Versao Origem"
            }
          },
          {
            "name": "versao_destino",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Versao Destino"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationComparison"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/calculos/{calculo_id}/estado": {
      "post": {
        "tags": [
          "Cálculos"
        ],
        "summary": "Change State",
        "description": "Arquiva, cancela ou reativa o cálculo sem excluir histórico.",
        "operationId": "change_state_api_v2_calculos__calculo_id__estado_post",
        "parameters": [
          {
            "name": "calculo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Calculo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculationStateChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationStateResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/calculos/{calculo_id}/versoes/{versao}/execucoes": {
      "get": {
        "tags": [
          "Cálculos"
        ],
        "summary": "Executions",
        "description": "Pagina execuções técnicas sem carregar indefinidamente uma versão muito reexecutada.",
        "operationId": "executions_api_v2_calculos__calculo_id__versoes__versao__execucoes_get",
        "parameters": [
          {
            "name": "calculo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Calculo Id"
            }
          },
          {
            "name": "versao",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Versao"
            }
          },
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Pagina"
            }
          },
          {
            "name": "tamanho_pagina",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Tamanho Pagina"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationExecutionsPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/calculos/{calculo_id}/versoes/{versao}": {
      "get": {
        "tags": [
          "Cálculos"
        ],
        "summary": "Version Detail",
        "description": "Reabre o snapshot exato de uma versão sem executar o motor novamente.",
        "operationId": "version_detail_api_v2_calculos__calculo_id__versoes__versao__get",
        "parameters": [
          {
            "name": "calculo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Calculo Id"
            }
          },
          {
            "name": "versao",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Versao"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationVersionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/calculos/{calculo_id}/versoes/{versao}/memoria-pdf": {
      "get": {
        "tags": [
          "Cálculos"
        ],
        "summary": "Version Pdf",
        "description": "Baixa a memória congelada quando aquela versão de negócio foi criada.",
        "operationId": "version_pdf_api_v2_calculos__calculo_id__versoes__versao__memoria_pdf_get",
        "parameters": [
          {
            "name": "calculo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Calculo Id"
            }
          },
          {
            "name": "versao",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Versao"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/calculos/{calculo_id}/execucoes/{execucao_id}/memoria-pdf": {
      "get": {
        "tags": [
          "Cálculos"
        ],
        "summary": "Execution Pdf",
        "description": "Baixa a memória da execução atual, mesmo quando a versão foi reutilizada.",
        "operationId": "execution_pdf_api_v2_calculos__calculo_id__execucoes__execucao_id__memoria_pdf_get",
        "parameters": [
          {
            "name": "calculo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Calculo Id"
            }
          },
          {
            "name": "execucao_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Execucao Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/calculos": {
      "post": {
        "tags": [
          "Cálculos"
        ],
        "summary": "Calculate",
        "description": "Calcula e versiona com controle otimista, inclusive a partir de versões históricas.",
        "operationId": "calculate_api_v2_calculos_post",
        "parameters": [
          {
            "name": "calculo_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calculo Id"
            }
          },
          {
            "name": "versao_base",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Versao Base"
            }
          },
          {
            "name": "versao_atual_esperada",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Versao Atual Esperada"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculationRequest-Input"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionedCalculationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/calculos/memoria-pdf": {
      "post": {
        "tags": [
          "Cálculos"
        ],
        "summary": "Memory Pdf",
        "description": "Exporta o request atual sem criar estado funcional ou execução persistida.",
        "operationId": "memory_pdf_api_v2_calculos_memoria_pdf_post",
        "parameters": [
          {
            "name": "indices_sha256",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Indices Sha256"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculationRequest-Input"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalculationDefaults": {
        "properties": {
          "mes": {
            "type": "string",
            "enum": [
              "janeiro",
              "fevereiro",
              "março",
              "abril",
              "maio",
              "junho",
              "julho",
              "agosto",
              "setembro",
              "outubro",
              "novembro",
              "dezembro"
            ],
            "title": "Mes"
          },
          "ano": {
            "type": "integer",
            "title": "Ano"
          },
          "competencia_recomendada": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competencia Recomendada"
          },
          "ajustada_por_disponibilidade": {
            "type": "boolean",
            "title": "Ajustada Por Disponibilidade",
            "default": false
          },
          "mensagem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mensagem"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mes",
          "ano"
        ],
        "title": "CalculationDefaults"
      },
...
      "CalculationPolicyView": {
        "properties": {
          "origem_calculo": {
            "type": "string",
            "enum": [
              "manual",
              "processo"
            ],
            "title": "Origem Calculo"
          },
          "parametros_padrao": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Parametros Padrao"
          },
          "campos_obrigatorios": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Campos Obrigatorios"
          },
          "catalogo": {
            "items": {
              "$ref": "#/components/schemas/ParameterCatalogItem"
            },
            "type": "array",
            "title": "Catalogo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "origem_calculo",
          "parametros_padrao",
          "campos_obrigatorios",
          "catalogo"
        ],
        "title": "CalculationPolicyView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}